pyproject.toml
[project]
name = "tldrify"
version = "0.1.0"
description = "Transform PDFs into intelligent study materials"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    # "surya-ocr>=0.4.0",  # Temporarily disabled for Python 3.13 compatibility
    "Pillow>=10.2.0",
    "pdf2image>=1.17.0",
    "PyPDF2>=3.0.1",
    "pymupdf>=1.23.8",  # Alternative OCR/PDF processing
    "sqlalchemy>=2.0.25",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.13.1",
    "redis>=5.0.1",
    "openai>=1.9.0",
    # "langchain>=0.1.0",  # Has complex dependencies
    # "tiktoken>=0.5.2",  # Not compatible with Python 3.13 yet
    "numpy>=1.26.3",
    "celery>=5.3.4",
    "flower>=2.0.1",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "qdrant-client>=1.7.0",  # Vector database
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "black>=23.12.1",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
    "ipython>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
target-version = "py311"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"